pyproject.toml
[project]
name = "web-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
